+++ app/shared/shared/services/field_vocab_mappings.py
@@ -1,7 +1,5 @@
 import logging
-from typing import List, Dict
-from collections import defaultdict
-
+from typing import List
 
 from shared.mapping.models import ScanReportTable, ScanReportField
 from shared.services.storage_service import StorageService
@@ -10,56 +8,6 @@ from shared.services.models import VocabularyMapping
 logger = logging.getLogger("test_table_mapping_logger")
 
 storage_service = StorageService()
-
-
-def validate_data_dictionary(data_dictionary: List[Dict]) -> None:
-    """
-    Validate the data dictionary for multiple empty 'value' mappings.
-
-    This function checks if there are multiple records in the
-    data dictionary with the same 'csv_file_name' and
-    'field_name' but empty 'value'.
-
-    If so, it raises an InvalidDataDictionaryError
-    because one field in one table can only
-    receive one vocabulary ID.
-
-    Args:
-        data_dictionary (List[Dict]): A list of dictionaries
-        representing the data dictionary.
-
-    Raises:
-        InvalidDataDictionaryError: If multiple records have
-        the same 'csv_file_name' and 'field_name'
-        with empty 'value'.
-
-    Returns:
-        None
-    """
-    # STEP 1: Group records by (csv_file_name, field_name)
-    field_groups = defaultdict(list)
-
-    # STEP 2: Iterate through the data dictionary and group records
-    for record in data_dictionary:
-        key = (record["csv_file_name"], record["field_name"])
-        field_groups[key].append(record)
-
-    # STEP 3:  Validate each group
-    for (csv_file, field_name), records in field_groups.items():
-
-        # Only check groups with >1 record
-        if len(records) > 1:
-
-            # Count records with empty 'value' (None or "")
-            empty_value_count = sum(1 for record in records if not record["value"])
-
-            # If >1 empty 'value' in the same group → ERROR
-            if empty_value_count > 1:
-                raise InvalidDataDictionaryError(
-                    f"Found {empty_value_count} empty 'value' mappings for "
-                    f"field '{field_name}' in file '{csv_file}'. "
-                    "One field in one table can only receive one vocab ID."
-                )
 
 
 def get_field_vocab_mappings(
@@ -128,11 +76,5 @@ def get_field_vocab_mappings(
 
 class GetFieldVocabMappingError(Exception):
     """Custom exception for errors in fetching field-vocab mappings."""
-
-    pass
-
-
-class InvalidDataDictionaryError(Exception):
-    """Custom exception for invalid data dictionary errors."""
 
     pass
